+++ frontend/src/homes/carbonEmission/Anomaly.jsx
@@ -7,73 +7,74 @@ import AnomalyBarChart from '@homes/carbonEmission/AnomalyBarChart.jsx';
 
 const Anomaly = () => {
 
-    // 공정 선택 상태 관리
-    const [activeStep, setActiveStep] = useState(null);
-    
-    // 모달 제어 상태 관리
-    const [selectedLog, setSelectedLog] = useState(null); 
-    const [isModalOpen, setIsModalOpen] = useState(false); 
-    const [modalStatus, setModalStatus] = useState("0");
-    const actionOptions = [
-        { value: "0", label: "미조치" },
-        { value: "1", label: "조치완료" }
-    ];
+  // 공정 선택 상태 관리
+  const [activeStep, setActiveStep] = useState(null);
 
-    // 상태 변경 모달 열기 함수
-    const openStatusModal = (e, log) => {
-        e.preventDefault();
-        setSelectedLog(log);
-        setModalStatus(log.isActioned ? "1" : "0"); 
-        setIsModalOpen(true);
-    };
-    // 상태 변경 모달 닫기 함수
-    const closeStatusModal = () => {
-        setSelectedLog(null);
-        setIsModalOpen(false);
-    };
+  // 모달 제어 상태 관리
+  const [selectedLog, setSelectedLog] = useState(null);
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [modalStatus, setModalStatus] = useState("0");
+  const actionOptions = [
+    { value: "0", label: "미조치" },
+    { value: "1", label: "조치완료" }
+  ];
 
-    // 상태 저장 함수
-    const saveStatus = () => {
-        if (!selectedLog) return;
-        const updatedIsActioned = modalStatus === "1"; 
+  // 상태 변경 모달 열기 함수
+  const openStatusModal = (e, log) => {
+    e.preventDefault();
+    setSelectedLog(log);
+    setModalStatus(log.isActioned ? "1" : "0");
+    setIsModalOpen(true);
+  };
+  // 상태 변경 모달 닫기 함수
+  const closeStatusModal = () => {
+    setSelectedLog(null);
+    setIsModalOpen(false);
+  };
 
-        setLogs(prevLogs =>
-        prevLogs.map(item =>
-            item.id === selectedLog.id ? { ...item, isActioned: updatedIsActioned } : item
-        )
-        );
-        closeStatusModal();
-    };
+  // 상태 저장 함수
+  const saveStatus = () => {
+    if (!selectedLog) return;
+    const updatedIsActioned = modalStatus === "1";
 
-    // 필터 상태 관리
-    const [selectedYear, setSelectedYear] = useState("all");
-    const [selectedQuarter, setSelectedQuarter] = useState("all");
-    const [selectedMonth, setSelectedMonth] = useState("none");
-    
-    // 분기 선택시 월 선택 초기화
-    const quarterChoice = (quarter) => {
+    setLogs(prevLogs =>
+      prevLogs.map(item =>
+        item.id === selectedLog.id ? { ...item, isActioned: updatedIsActioned } : item
+      )
+    );
+    closeStatusModal();
+  };
+
+  // 필터 상태 관리
+  const [selectedYear, setSelectedYear] = useState("all");
+  const [selectedQuarter, setSelectedQuarter] = useState("all");
+  const [selectedMonth, setSelectedMonth] = useState("none");
+
+  // 분기 선택시 월 선택 초기화
+  const quarterChoice = (quarter) => {
     setSelectedQuarter(quarter);
     setSelectedMonth("none");
-    };
-    // 월 선택시 분기 선택 초기화
-    const monthChoice = (e) => {
-        const month = e.target.value;
-        setSelectedMonth(month);
-        if (month !== "none") {
-        setSelectedQuarter("none");
-        }
-        };
-    
-    // 조회 버튼 이벤트
-    const periodSearch = () => {
-        const filterPayload = {
-        year: selectedYear,
-        quarter: selectedQuarter,
-        month: selectedMonth
-        }};
+  };
+  // 월 선택시 분기 선택 초기화
+  const monthChoice = (e) => {
+    const month = e.target.value;
+    setSelectedMonth(month);
+    if (month !== "none") {
+      setSelectedQuarter("none");
+    }
+  };
 
-    // STEP 카드 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
-    const stepData = [
+  // 조회 버튼 이벤트
+  const periodSearch = () => {
+    const filterPayload = {
+      year: selectedYear,
+      quarter: selectedQuarter,
+      month: selectedMonth
+    }
+  };
+
+  // STEP 카드 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
+  const stepData = [
     {
       id: 1,
       badge: "STEP 1",
@@ -149,7 +150,7 @@ const Anomaly = () => {
       indirectEmission: "13.3272 tCO₂e",
       currentEmission: "15.1180 tCO₂e",
       directEmission: "13.3272 tCO₂e",
-      directCurrentEmission:"13.3000 tCO₂e",
+      directCurrentEmission: "13.3000 tCO₂e",
       anomalyCount: "1건 (조치완료)",
       countHighlight: true
     },
@@ -164,150 +165,150 @@ const Anomaly = () => {
       anomalyCount: "1건 (조치완료)",
       countHighlight: false
     }
-        ];
+  ];
 
-    // 상세 정보 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
-    const detailsData = [
-            {
-            id: 1,
-            title: "STEP 1 · 빌릿 가열 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "147.03",
-                actual: "139.8389",
-                deviation: "-4.89%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            },
-            {
-            id: 2,
-            title: "STEP 2 · 간접 압출 공정",
-            headerClass: "text-critical",
-            icon: "⚠",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "Critical",
-                isCritical: true
-                }
-            ]
-            },
-            {
-            id: 3, 
-            title: "STEP 3 · 인발 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "147.03",
-                actual: "139.8389",
-                deviation: "-4.89%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            },
-            {
-            id: 4, 
-            title: "STEP 4 · 알루미늄 시효 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            },
-            {
-            id: 5, 
-            title: "STEP 5 · 4축 CNC 가공 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            },
-            {
-            id: 6, 
-            title: "STEP 6 · 알루미늄 절단 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            },
-            {
-            id: 7, 
-            title: "STEP 7 · 아노다이징 공정",
-            headerClass: "text-critical",
-            icon: "⚠",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "critical",
-                isCritical: true
-                }
-            ]
-            },
-            {
-            id: 8, 
-            title: "STEP 8 · 구리스 디스펜싱 공정",
-            headerClass: "text-success",
-            icon: "✓",
-            metrics: [
-                {
-                name: "전력 사용량 (kWh)",
-                standard: "73.51",
-                actual: "121.001",
-                deviation: "+64.60%",
-                status: "정상",
-                isCritical: false
-                }
-            ]
-            }
-        ];
-    // 상세 정보 열기 함수
-    const pipelineStepClick = (step) => {
-        setActiveStep(prevStep => prevStep === step ? null : step);
-        };
-    // 상세 정보 닫기 함수
-    const closeDetail = () => {
-                    setActiveStep(null);
-                };
+  // 상세 정보 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
+  const detailsData = [
+    {
+      id: 1,
+      title: "STEP 1 · 빌릿 가열 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "147.03",
+          actual: "139.8389",
+          deviation: "-4.89%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    },
+    {
+      id: 2,
+      title: "STEP 2 · 간접 압출 공정",
+      headerClass: "text-critical",
+      icon: "⚠",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "Critical",
+          isCritical: true
+        }
+      ]
+    },
+    {
+      id: 3,
+      title: "STEP 3 · 인발 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "147.03",
+          actual: "139.8389",
+          deviation: "-4.89%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    },
+    {
+      id: 4,
+      title: "STEP 4 · 알루미늄 시효 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    },
+    {
+      id: 5,
+      title: "STEP 5 · 4축 CNC 가공 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    },
+    {
+      id: 6,
+      title: "STEP 6 · 알루미늄 절단 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    },
+    {
+      id: 7,
+      title: "STEP 7 · 아노다이징 공정",
+      headerClass: "text-critical",
+      icon: "⚠",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "critical",
+          isCritical: true
+        }
+      ]
+    },
+    {
+      id: 8,
+      title: "STEP 8 · 구리스 디스펜싱 공정",
+      headerClass: "text-success",
+      icon: "✓",
+      metrics: [
+        {
+          name: "전력 사용량 (kWh)",
+          standard: "73.51",
+          actual: "121.001",
+          deviation: "+64.60%",
+          status: "정상",
+          isCritical: false
+        }
+      ]
+    }
+  ];
+  // 상세 정보 열기 함수
+  const pipelineStepClick = (step) => {
+    setActiveStep(prevStep => prevStep === step ? null : step);
+  };
+  // 상세 정보 닫기 함수
+  const closeDetail = () => {
+    setActiveStep(null);
+  };
 
-    // 로그 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
-    const [logs, setLogs] = useState([
+  // 로그 데이터 예시 (실제 데이터는 API 연동 후 동적으로 관리)
+  const [logs, setLogs] = useState([
     {
       id: 1,
       date: "2026-06-05 14:15:22",
@@ -332,15 +333,15 @@ const Anomaly = () => {
 
 
   return (
-   <>
-    <div className="main-wrapper">
+    <>
+
 
       {/* 본문 영역 */}
-      <div className="content-area">
+      
         <div className="page-header-flex">
           <h2 className="page-title">이상치 탐지</h2>
-        </div>   
-        
+        </div>
+
         <div className="pipeline-section-wrapper">
           <div className="pipeline-header-flex">
             <h2>작일 공정 파이프라인 이상치 현황</h2>
@@ -348,187 +349,187 @@ const Anomaly = () => {
 
           <div className="pipeline-flex-container">
             {stepData.map((step, index) => (
-                <React.Fragment key={step.id}>
+              <React.Fragment key={step.id}>
                 {/* STEP 카드 영역 */}
-                <div 
-                    className={`pipeline-step-card ${step.isCritical ? 'critical' : ''} ${activeStep === step.id ? 'active' : ''}`} 
-                    onClick={() => pipelineStepClick(step.id)}
+                <div
+                  className={`pipeline-step-card ${step.isCritical ? 'critical' : ''} ${activeStep === step.id ? 'active' : ''}`}
+                  onClick={() => pipelineStepClick(step.id)}
                 >
-                    <div className="step-header">
+                  <div className="step-header">
                     <span className="step-badge">{step.badge}</span>
                     <span className="step-title">
-                        <span className={`dot ${step.dotClass}`}></span> {step.title}
+                      <span className={`dot ${step.dotClass}`}></span> {step.title}
                     </span>
-                    </div>
-                    <div className="step-info-row">
+                  </div>
+                  <div className="step-info-row">
                     <span className="label">적정 간접 탄소 배출량</span>
                     <span className={`value ${step.id === 2 ? 'highlight' : ''}`}>{step.indirectEmission}</span>
-                    </div>
-                    <div className="step-info-row">
+                  </div>
+                  <div className="step-info-row">
                     <span className="label">간접 탄소 배출량</span>
                     <span className="value">{step.currentEmission}</span>
-                    </div>
-                    {step.directEmission && ( 
+                  </div>
+                  {step.directEmission && (
                     <>
-                    <div className="step-info-row">
-                    <span className="label">적정 직접 탄소 배출량</span>
-                    <span className="value">{step.directEmission}</span>
-                    </div>
-                    <div className="step-info-row">
-                    <span className="label">직접 탄소 배출량</span>
-                    <span className="value">{step.directCurrentEmission}</span>
-                    </div>
+                      <div className="step-info-row">
+                        <span className="label">적정 직접 탄소 배출량</span>
+                        <span className="value">{step.directEmission}</span>
+                      </div>
+                      <div className="step-info-row">
+                        <span className="label">직접 탄소 배출량</span>
+                        <span className="value">{step.directCurrentEmission}</span>
+                      </div>
                     </>
-                )}
-                    <div className="step-info-row">
+                  )}
+                  <div className="step-info-row">
                     <span className="label">이상치 건수</span>
                     <span className={`value ${step.countHighlight ? 'highlight' : ''}`}>{step.anomalyCount}</span>
-                    </div>
+                  </div>
                 </div>
                 {index < stepData.length - 1 && <div className="arrow-icon">➔</div>}
-                </React.Fragment>
+              </React.Fragment>
             ))}
-            </div>
+          </div>
 
-            {detailsData.map((detail) => {
+          {detailsData.map((detail) => {
             if (activeStep !== detail.id) return null;
             return (
-            <div className="pipeline-detail-box" key={detail.id}>
+              <div className="pipeline-detail-box" key={detail.id}>
                 {/* 헤더 영역 */}
                 <div className={`detail-header ${detail.headerClass}`}>
-                <span>{detail.icon} {detail.title}</span>
-                <button type="button" className="close-btn" onClick={(e) => {
+                  <span>{detail.icon} {detail.title}</span>
+                  <button type="button" className="close-btn" onClick={(e) => {
                     e.stopPropagation();
                     setActiveStep(null);
-                }}>✕</button>
+                  }}>✕</button>
                 </div>
 
                 {/* 본문 테이블 영역 */}
                 <div className="detail-content">
-                <table className="status-table">
+                  <table className="status-table">
                     <thead>
-                    <tr>
+                      <tr>
                         <th>측정 항목</th>
                         <th>기준치 (표준)</th>
                         <th>실측값</th>
                         <th>편차율</th>
                         <th>상태</th>
-                    </tr>
+                      </tr>
                     </thead>
                     <tbody>
-                    {detail.metrics.map((metric, index) => (
+                      {detail.metrics.map((metric, index) => (
                         <tr key={index} className={metric.isCritical ? "row-critical" : ""}>
-                        <td>
+                          <td>
                             {metric.isCritical ? <strong>{metric.name}</strong> : metric.name}
-                        </td>
-                        <td>{metric.standard}</td>
-                        <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
+                          </td>
+                          <td>{metric.standard}</td>
+                          <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
                             {metric.actual}
-                        </td>
-                        <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
+                          </td>
+                          <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
                             {metric.deviation}
-                        </td>
-                        <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
+                          </td>
+                          <td className={`status-text ${metric.isCritical ? "critical" : "normal"}`}>
                             {metric.status}
-                        </td>
+                          </td>
                         </tr>
-                    ))}
+                      ))}
                     </tbody>
-                </table>
+                  </table>
                 </div>
-            </div>
+              </div>
             );
-      })}
+          })}
 
           <h2 className="period-section-title">기간 별 공정 이상치 현황</h2>
         </div>
 
         {/* 필터 바 */}
-            <div className="dashboard-filter-bar">
-            <div className="toggle-group-wrapper">
+        <div className="dashboard-filter-bar">
+          <div className="toggle-group-wrapper">
 
-                {/* 연도 선택 */}
-                <div className="toggle-group-item id_quarter">
-                <div className="select_label"><span>연도 선택</span></div>
-                <div className="toggle-content">
-                    <select 
-                    className="year-select btn-filter-tab" 
-                    value={selectedYear}
-                    onChange={(e) => {
-                        setSelectedYear(e.target.value);
-                        setSelectedQuarter("all");
-                        setSelectedMonth("none");
-                    }}
-                    >
-                    <option value="all">전체</option>
-                    <option value="2026">2026년</option>
-                    <option value="2025">2025년</option>
-                    <option value="2024">2024년</option>
-                    <option value="2023">2023년</option>
-                    </select>
-                </div>
-                </div>
-
-                {selectedYear !== "all" && (
-                <>
-                    {/* 분기 선택 영역 */}
-                    <div className="toggle-group-item id_quarter">
-                    <div className="select_label"><span>분기 선택</span></div>
-                    <div className="toggle-content">
-                        {[
-                        { code: "all", name: "전체" },
-                        { code: "1", name: "1분기" },
-                        { code: "2", name: "2분기" },
-                        { code: "3", name: "3분기" },
-                        { code: "4", name: "4분기" }
-                        ].map((q) => (
-                        <button 
-                            key={q.code}
-                            type="button" 
-                            className={`btn-filter-tab ${selectedQuarter === q.code ? 'active' : ''}`}
-                            onClick={() => quarterChoice(q.code)}
-                        >
-                            {q.name}
-                        </button>
-                        ))}
-                    </div>
-                    </div>
-
-                    {/* 월 선택 영역 */}
-                    <div className="toggle-group-item id_quarter">
-                    <div className="select_label"><span>월 선택</span></div>
-                    <div className="toggle-content">
-                        <select 
-                        className="year-select btn-filter-tab" 
-                        value={selectedMonth}
-                        onChange={monthChoice}
-                        >
-                        <option value="none">선택</option>
-                        {Array.from({ length: 12 }, (_, i) => i + 1).map(month => (
-                            <option key={month} value={month}>{month}월</option>
-                        ))}
-                        </select>
-                    </div>
-                    </div>
-                </>
-                )}
-            <div className="toggle-group-item" style={{ marginLeft: 'auto' }}>
-                <button 
-                    type="button" 
-                    className="btn-filter-tab active"
-                    onClick={periodSearch}>
-                    조회
-                </button>
-                </div>
+            {/* 연도 선택 */}
+            <div className="toggle-group-item id_quarter">
+              <div className="select_label"><span>연도 선택</span></div>
+              <div className="toggle-content">
+                <select
+                  className="year-select btn-filter-tab"
+                  value={selectedYear}
+                  onChange={(e) => {
+                    setSelectedYear(e.target.value);
+                    setSelectedQuarter("all");
+                    setSelectedMonth("none");
+                  }}
+                >
+                  <option value="all">전체</option>
+                  <option value="2026">2026년</option>
+                  <option value="2025">2025년</option>
+                  <option value="2024">2024년</option>
+                  <option value="2023">2023년</option>
+                </select>
+              </div>
             </div>
-            </div> 
+
+            {selectedYear !== "all" && (
+              <>
+                {/* 분기 선택 영역 */}
+                <div className="toggle-group-item id_quarter">
+                  <div className="select_label"><span>분기 선택</span></div>
+                  <div className="toggle-content">
+                    {[
+                      { code: "all", name: "전체" },
+                      { code: "1", name: "1분기" },
+                      { code: "2", name: "2분기" },
+                      { code: "3", name: "3분기" },
+                      { code: "4", name: "4분기" }
+                    ].map((q) => (
+                      <button
+                        key={q.code}
+                        type="button"
+                        className={`btn-filter-tab ${selectedQuarter === q.code ? 'active' : ''}`}
+                        onClick={() => quarterChoice(q.code)}
+                      >
+                        {q.name}
+                      </button>
+                    ))}
+                  </div>
+                </div>
+
+                {/* 월 선택 영역 */}
+                <div className="toggle-group-item id_quarter">
+                  <div className="select_label"><span>월 선택</span></div>
+                  <div className="toggle-content">
+                    <select
+                      className="year-select btn-filter-tab"
+                      value={selectedMonth}
+                      onChange={monthChoice}
+                    >
+                      <option value="none">선택</option>
+                      {Array.from({ length: 12 }, (_, i) => i + 1).map(month => (
+                        <option key={month} value={month}>{month}월</option>
+                      ))}
+                    </select>
+                  </div>
+                </div>
+              </>
+            )}
+            <div className="toggle-group-item" style={{ marginLeft: 'auto' }}>
+              <button
+                type="button"
+                className="btn-filter-tab active"
+                onClick={periodSearch}>
+                조회
+              </button>
+            </div>
+          </div>
+        </div>
 
         {/* 차트 영역 */}
         <div style={{ marginBottom: '40px' }}>
           <div className="chart-card">
             <h3 className="chart-title">설비별 이상치 발생 횟수 현황</h3>
             <div className="chart-container chart-height-machine">
-                <AnomalyBarChart dataValues={[4, 12, 2, 5, 8, 1, 0, 3]} />
+              <AnomalyBarChart dataValues={[4, 12, 2, 5, 8, 1, 0, 3]} />
             </div>
           </div>
         </div>
@@ -551,35 +552,35 @@ const Anomaly = () => {
                   <th>조치 상태</th>
                 </tr>
               </thead>
-             <tbody>
+              <tbody>
                 {logs.map((log) => (
-                    <tr key={log.id}>
+                  <tr key={log.id}>
                     <td className="text-center">{log.date}</td>
                     <td className="font-bold text-center">{log.machineName}</td>
                     <td className="text-center font-bold">{log.scope}</td>
                     <td className="text-center">{log.metrics}</td>
                     <td className="text-center">
-                        <span style={{ color: log.levelColor, fontWeight: 600 }}>{log.level}</span>
+                      <span style={{ color: log.levelColor, fontWeight: 600 }}>{log.level}</span>
                     </td>
                     <td className="text-center">
-                        <button 
+                      <button
                         onClick={(e) => openStatusModal(e, log)}
-                        className={`btn-filter-tab ${!log.isActioned ? "active" : ""}`} 
-                        style={{ 
-                            padding: '2px 6px', 
-                            fontSize: '11px', 
-                            backgroundColor: !log.isActioned ? '#e53e3e' : 'transparent',
-                            borderColor: !log.isActioned ? '#e53e3e' : '#ccc',
-                            color: !log.isActioned ? '#fff' : '#333',
-                            cursor: 'pointer'
+                        className={`btn-filter-tab ${!log.isActioned ? "active" : ""}`}
+                        style={{
+                          padding: '2px 6px',
+                          fontSize: '11px',
+                          backgroundColor: !log.isActioned ? '#e53e3e' : 'transparent',
+                          borderColor: !log.isActioned ? '#e53e3e' : '#ccc',
+                          color: !log.isActioned ? '#fff' : '#333',
+                          cursor: 'pointer'
                         }}
-                        >
+                      >
                         {log.isActioned ? actionOptions[1].label : actionOptions[0].label}
-                        </button>
+                      </button>
                     </td>
-                    </tr>
+                  </tr>
                 ))}
-                </tbody>
+              </tbody>
               <tfoot>
                 <tr className="table-summary-row">
                   <td colSpan={3} className="text-center">선택기간 총 이상 발생 건수</td>
@@ -590,50 +591,50 @@ const Anomaly = () => {
               </tfoot>
             </table>
             {isModalOpen && selectedLog && (
-        <div id="anomalyStatusModal" className="anomaly-modal" style={{ display: 'flex' }}>
-          <div className="anomaly-modal-content">
-            <div className="anomaly-modal-header">조치 상태 변경</div>
-            <div className="anomaly-modal-body">
-              <select 
-                id="statusSelect" 
-                className="anomaly-select-control" 
-                value={modalStatus}
-                onChange={(e) => setModalStatus(e.target.value)}
-              >
-                <option value="0">미조치</option>
-                <option value="1">조치완료</option>
-              </select>
-            </div>
-            <div className="anomaly-modal-footer">
-              <button type="button" className="btn-modal-action" onClick={closeStatusModal}>취소</button>
-              <button type="button" className="btn-modal-action save" onClick={saveStatus}>저장</button>
-            </div>
-          </div>
-        </div>
-      )}
-      
-    </div>
+              <div id="anomalyStatusModal" className="anomaly-modal" style={{ display: 'flex' }}>
+                <div className="anomaly-modal-content">
+                  <div className="anomaly-modal-header">조치 상태 변경</div>
+                  <div className="anomaly-modal-body">
+                    <select
+                      id="statusSelect"
+                      className="anomaly-select-control"
+                      value={modalStatus}
+                      onChange={(e) => setModalStatus(e.target.value)}
+                    >
+                      <option value="0">미조치</option>
+                      <option value="1">조치완료</option>
+                    </select>
+                  </div>
+                  <div className="anomaly-modal-footer">
+                    <button type="button" className="btn-modal-action" onClick={closeStatusModal}>취소</button>
+                    <button type="button" className="btn-modal-action save" onClick={saveStatus}>저장</button>
+                  </div>
+                </div>
+              </div>
+            )}
+
+
           </div>
         </div>
         {/* 페이지네이션 */}
-    </div>
-          <div className="pagination-info">
-            전체 <span>74</span>건
-          </div>
-        <div className="pagination-container">
-          <div className="pagination-buttons">
-            <button type="button" className="btn-page first" title="첫 페이지" disabled>&laquo;</button>
-            <button type="button" className="btn-page prev" title="이전 페이지" disabled>&lsaquo;</button>
-            <button type="button" className="btn-page-num active">1</button>
-            <button type="button" className="btn-page-num">2</button>
-            <button type="button" className="btn-page-num">3</button>
-            <button type="button" className="btn-page-num">4</button>
-            <button type="button" className="btn-page-num">5</button>
-            <button type="button" className="btn-page next" title="다음 페이지">&rsaquo;</button>
-            <button type="button" className="btn-page last" title="마지막 페이지">&raquo;</button>
+      
+      <div className="pagination-info">
+        전체 <span>74</span>건
+      </div>
+      <div className="pagination-container">
+        <div className="pagination-buttons">
+          <button type="button" className="btn-page first" title="첫 페이지" disabled>&laquo;</button>
+          <button type="button" className="btn-page prev" title="이전 페이지" disabled>&lsaquo;</button>
+          <button type="button" className="btn-page-num active">1</button>
+          <button type="button" className="btn-page-num">2</button>
+          <button type="button" className="btn-page-num">3</button>
+          <button type="button" className="btn-page-num">4</button>
+          <button type="button" className="btn-page-num">5</button>
+          <button type="button" className="btn-page next" title="다음 페이지">&rsaquo;</button>
+          <button type="button" className="btn-page last" title="마지막 페이지">&raquo;</button>
         </div>
       </div>
-  </> 
+    </>
   )
 };
