+++ frontend/src/homes/main/index.jsx
@@ -5,11 +5,11 @@ import Sidebar from '@components/Layout/Sidebar.jsx';
 import Main from '@homes/Main/Main.jsx';
 import Asn from '@homes/inbound/Asn.jsx';
 import Inbound from '@homes/inbound/Inbound.jsx';
+import OutboundOrder from '@homes/outbound/OutboundOrder.jsx';
 import Outbound from '@homes/outbound/Outbound.jsx';
 import OutHistory from '@homes/outbound/OutHistory.jsx';
 import Anomaly from '@homes/carbonEmission/Anomaly.jsx';
 import CarbonDashboard from '@homes/carbonEmission/CarbonDashboard.jsx';
-import OutboundOrder from '@homes/outboundOP/OutboundOrder.jsx';
 
 
 const Home = () => {
@@ -23,11 +23,11 @@ const Home = () => {
               <Route path='/' element={<Main />}/>
               <Route path='/asn' element={<Asn />}/>
               <Route path='/inbound' element={<Inbound />}/>
+              <Route path='/order' element={<OutboundOrder />}/>
               <Route path='/outbound' element={<Outbound />}/>
               <Route path='/outhistory' element={<OutHistory />}/>
               <Route path='/anomaly' element={<Anomaly />}/>
               <Route path='/carbonemission' element={<CarbonDashboard />}/>
-              <Route path='/order' element={<OutboundOrder />}/>
             </Routes>
           </div>
       </div>
